requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
coverage==7.4.1